+++ libs/render_delegate/utils.cpp
@@ -110,7 +110,7 @@ void HdArnoldSetTransform(AtNode* node, HdSceneDelegate* sceneDelegate, const Sd
         xf.values.resize(transformKeys);
         // If an amount of transform keys is provided, we must resample
         // the times & values to match the new amount
-        for (size_t i = 0; i < transformKeys; ++i) {
+        for (int i = 0; i < transformKeys; ++i) {
             xf.times[i] = timeStart + i * (timeEnd - timeStart) /
                 (static_cast<float>(transformKeys)-1.f);
             xf.values[i] = xfOrig.Resample(xf.times[i]);
